lint, clean and format

+++ test/e2e/tests/consensus_test.go
@@ -30,7 +30,7 @@ func TestEvidence_DoubleVote(t *testing.T) {
 		require.NoError(t, err)
 
 		for height, behavior := range behaviors {
-			if behavior.String == "double-prevote" && int64(height) < status.SyncInfo.LatestBlockHeight {
+			if behavior.String == "double-prevote" && height < status.SyncInfo.LatestBlockHeight {
 				// we expect evidence to be formed in the height directly after hence height + 1
 				var reportHeight int64 = height + 1
 				resp, err := client.Block(ctx, &reportHeight)
@@ -38,6 +38,8 @@ func TestEvidence_DoubleVote(t *testing.T) {
 				assert.NotEmpty(t, resp.Block.Evidence.Evidence, "no evidence seen of node %v equivocating at height %d",
 					node.Name, height)
 				containsMaverick := false
+				// we expect that at least one of the evidence committed at this height will be duplicate vote evidence
+				// and contain the maverick nodes address
 				for _, ev := range resp.Block.Evidence.Evidence {
 					if dev, ok := ev.(*types.DuplicateVoteEvidence); ok {
 						if bytes.Equal(dev.VoteA.ValidatorAddress, node.Key.PubKey().Address()) {
